Removed another circular reference
