+++ modules/util.py
@@ -374,13 +374,6 @@ def ordinal_suffix(number: int) -> str:
     return 'th' if 10 <= number % 100 <= 20 else {1: 'st', 2: 'nd', 3: 'rd'}.get(number % 10, 'th')
 
 
-
-def makedirs_with_log(path):
-    try:
-        os.makedirs(path, exist_ok=True)
-    except OSError as error:
-        print(f'Directory {path} could not be created, reason: {error}')
-
 def get_enabled_loras(loras: list) -> list:
     return [[lora[1], lora[2]] for lora in loras if lora[0]]
 
@@ -402,3 +395,4 @@ def parse_lora_references_from_prompt(prompt: str, loras: List[Tuple[AnyStr, flo
             updated_loras.append(lora)
 
     return updated_loras[:loras_limit]
+
